+++ NEWS.md
@@ -1,4 +1,4 @@
-# random.cdisc.data 0.3.13.9012
+# random.cdisc.data 0.3.13.9013
 
 ### New features
 * Added new random dataset generator: Anti-Drug Antibody Analysis Dataset (`radab`).
@@ -12,6 +12,8 @@
 * Enhanced `radpp` with additional default `PARAM`s.
 * Add new variables `PPSTINT` `PPENINT` `REGIMEN` `PKARMCD` to `radpp`.
 * Enhanced `radab` variables `ADAPBLFL`, `ABLFL`, `ADABLFL`, `PARCAT1`, `RELTMU`
+* Add new variable `lgTMATRSK` to `adtte`.
+* Add `TNE` to `PARAMCD` and `Total Number of Exacerbations` to `PARAM` in `adtte`.
 
 # random.cdisc.data 0.3.13